requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
exceptiongroup==1.0.4
flake8==6.0.0
h11==0.14.0
idna==3.4
mccabe==0.7.0
mysql-connector-python==8.0.31
outcome==1.2.0
protobuf==3.20.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PySocks==1.7.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
